pyproject.toml
[project]
name = "ffmpeg-voice-vox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.9"
dependencies = [
    "black>=25.12.0",
    "faster-whisper>=1.2.1",
    "flask>=3.1.2",
    "numpy>=2.4.1",
    "psutil>=7.2.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.7.2",
    "pytest-watch>=4.2.0",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
]
